third_party/blink/web_tests/crypto/subtle/aes-cbc/unwrap-rsa-expected.txt

Test unwrapping an RSA key with AES-CBC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing an unwrapping key...
PASS unwrappingKey.algorithm.name is 'AES-CBC'
Unwrapping a key...
PASS unwrappedKey.toString() is '[object CryptoKey]'
PASS unwrappedKey.type is 'public'
PASS unwrappedKey.usages is ['verify']
PASS unwrappedKey.algorithm.name is 'RSASSA-PKCS1-v1_5'
PASS unwrappedKey.algorithm.modulusLength is 2048
PASS Base64URL.stringify(unwrappedKey.algorithm.publicExponent) is publicKeyJSON.e
PASS unwrappedKey.algorithm.hash.name is 'SHA-256'
PASS unwrappedKey.extractable is false
PASS successfullyParsed is true

TEST COMPLETE
